Tear apart the soundcloud app and repurpose
The majority of the effort has been removing the unnecessary
indirection in the android implementation.

I have been able to show that both android & iOS can stream from https
and also from local mp3 files. However getting the paths to those
local file is an arse on iOS we dont have a nice way to do this in
Fuse yet. So it's time to make a MediaQuery library which should help
out with this.